Modified StatsController.class name
Modified StatsController.class name to correct a typo (StatsContoller)
